Added neural net code.
Below are brief descriptions of the classes. Each header file has comments for each function and variable used. The implementation files have comments explaining parts of the code, but most of it you can ignore if you don't want to look at it.

Main - just testing and training code for now, will be removed in final version
C

Skynet - Class containing the net itself and the functions to build it, train it, and show the results.

Layer - Just a #define statement to eliminate a circular dependency and make my IDE stop flagging errors.

Neuron - the neurons of the net. These contain the data of the net for input, training, and output.